pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "meeting-transcription"
version = "1.0.0"
description = "Production-grade meeting transcription with speaker diarization, identification, and Whisper transcription."
requires-python = ">=3.10"
readme = "README.md"
license = { text = "MIT" }

dependencies = [
    "torch>=2.1",
    "torchaudio>=2.1",
    "soundfile>=0.12",
    "resampy>=0.4",
    "demucs>=4.0",
    "pyannote.audio>=3.1",
    "openai-whisper>=20231117",
    "resemblyzer>=0.1.1",
    "gradio>=4.20",
    "numpy>=1.24",
    "python-dotenv>=1.0",
]

[project.scripts]
meeting-transcription = "run:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"
